Use webpack bundled chart library assets
