+++ Resources/Public/JavaScript/apexcharts.js
@@ -1,287 +1,1 @@
-var Hoogi91 = Hoogi91 || {};
-Hoogi91.Charts = {
-    chartsData: null,
-    chartsContainer: null,
-
-    /**
-     * initialize all charts that can be found if available options are filled
-     */
-    init: function () {
-        if (typeof Hoogi91.chartsData === 'object') {
-            this.chartsData = Hoogi91.chartsData;
-        }
-
-        this.chartsContainer = document.querySelectorAll('.chart-container > .chart');
-        if (this.chartsContainer.length > 0 && this.chartsData !== null && Object.keys(this.chartsData).length > 0) {
-            for (var i = 0; i < this.chartsContainer.length; ++i) {
-                var element = this.chartsContainer[i];
-
-                // check id, type and datakey of current chart element and break if something is not available
-                var type = element.getAttribute('data-chart-type') || '';
-                var dataKey = element.getAttribute('data-chart-data') || '';
-                if (type.length <= 0 || dataKey.length <= 0) {
-                    continue;
-                }
-
-                var chartData = this.getChartData(dataKey);
-                if (chartData.length <= 0) {
-                    continue;
-                }
-
-                var labels = chartData.labels || [];
-                var datasets = chartData.datasets || [];
-                if (labels.length > 0 && datasets.length > 0) {
-                    switch (type) {
-                        case 'chart_bar':
-                            this.createBarChart(element, labels, datasets);
-                            break;
-                        case 'chart_line':
-                            this.createLineChart(element, labels, datasets);
-                            break;
-                        case 'chart_pie':
-                            this.createPieChart(element, labels, datasets);
-                            break;
-                        case 'chart_doughnut':
-                            this.createDoughnutChart(element, labels, datasets);
-                            break;
-                    }
-                }
-            }
-        }
-    },
-
-    /**
-     * get chart data by key
-     *
-     * @param key
-     * @returns object
-     */
-    getChartData: function (key) {
-        if (typeof this.chartsData[key] === 'undefined') {
-            return {};
-        }
-        return this.chartsData[key];
-    },
-
-    /**
-     * build bar chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createBarChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        var chart = new ApexCharts(element, {
-            chart: {
-                type: 'bar',
-                stacked: _this.getKeyOfObject(options, 'bar.stacked', 0) === '1',
-            },
-            series: _this.createDatasets(datasets, function (set) {
-                var dataset = _this.createDatasets(set['data'], function (value, index) {
-                    var dataRow = {x: labels[index] || '', y: value};
-                    if (datasets.length === 1 && set['background'].length > 1) {
-                        dataRow['fillColor'] = set['background'][index] || null;
-                    }
-                    if (datasets.length === 1 && set['border'].length > 1) {
-                        dataRow['strokeColor'] = set['border'][index] || null;
-                    }
-                    return dataRow;
-                });
-                return {name: set['label'], data: dataset}
-            }),
-            labels: labels,
-            colors: datasets[0]['border'] || datasets[0]['background'] || undefined,
-            fill: {
-                colors: datasets[0]['background'] || datasets[0]['border'] || undefined,
-            },
-            dataLabels: {enabled: false},
-            stroke: {width: 1},
-            plotOptions: {
-                bar: {
-                    horizontal: _this.getKeyOfObject(options, 'bar.horizontal', 0) === '1',
-                }
-            },
-            xaxis: _this.getTicksConfig(options, 'x'),
-            yaxis: _this.getTicksConfig(options, 'y'),
-            legend: {
-                show: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                position: _this.getKeyOfObject(options, 'legend.position', 'top'),
-            },
-        });
-        chart.render();
-
-        return chart;
-    },
-
-    /**
-     * build line chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createLineChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        var chart = new ApexCharts(element, {
-            chart: {
-                type: _this.getKeyOfObject(options, 'line.fill', 0) === '1' ? 'area' : 'line',
-                stacked: _this.getKeyOfObject(options, 'line.stacked', 0) === '1',
-            },
-            series: _this.createDatasets(datasets, function (set) {
-                return {name: set['label'], data: set['data']};
-            }),
-            labels: labels,
-            colors: datasets[0]['border'] || datasets[0]['background'] || undefined,
-            fill: {
-                colors: datasets[0]['background'] || datasets[0]['border'] || undefined,
-            },
-            dataLabels: {enabled: false},
-            stroke: {
-                colors: datasets[0]['background'] || undefined,
-                curve: _this.getKeyOfObject(options, 'line.stepped', 'smooth'),
-            },
-            markers: {
-                size: [5, 7],
-                colors: datasets[0]['border'] || undefined,
-            },
-            xaxis: _this.getTicksConfig(options, 'x'),
-            yaxis: _this.getTicksConfig(options, 'y'),
-            legend: {
-                show: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                position: _this.getKeyOfObject(options, 'legend.position', 'top'),
-            },
-        });
-        chart.render();
-
-        return chart;
-    },
-
-    /**
-     * build pie chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createPieChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        console.log(labels)
-        var chart = new ApexCharts(element, {
-            chart: {type: 'pie'},
-            colors: datasets[0]['background'] || datasets[0]['border'] || [],
-            series: datasets[0]['data'] || [],
-            labels: labels,
-            dataLabels: {enabled: false},
-            legend: {
-                show: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                position: _this.getKeyOfObject(options, 'legend.position', 'top'),
-            },
-        });
-        chart.render();
-
-        return chart;
-    },
-
-    /**
-     * build doughnut chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createDoughnutChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        var cutoutValue = parseInt(_this.getKeyOfObject(options, 'doughnut.cutoutPercentage', 75), 10);
-        var chart = new ApexCharts(element, {
-            chart: {type: 'donut'},
-            colors: datasets[0]['background'] || datasets[0]['border'] || [],
-            series: datasets[0]['data'] || [],
-            labels: labels,
-            dataLabels: {enabled: false},
-            plotOptions: {
-                pie: {donut: {size: cutoutValue <= 0 ? 75 : cutoutValue}}
-            },
-            legend: {
-                show: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                position: _this.getKeyOfObject(options, 'legend.position', 'top'),
-            },
-        });
-        chart.render();
-
-        return chart;
-    },
-
-    /**
-     * create datasets by given data and an additional mapping callback
-     *
-     * @param datasets
-     * @param mapping
-     * @returns array
-     */
-    createDatasets: function (datasets, mapping) {
-        var keyValueMapping = mapping || null,
-            processedDatasets = [];
-
-        if (typeof keyValueMapping !== 'function') {
-            return processedDatasets;
-        }
-
-        if (typeof datasets === 'object' && datasets.length > 0) {
-            for (var i = 0; i < datasets.length; ++i) {
-                var set = keyValueMapping(datasets[i], i);
-                processedDatasets.push(set);
-            }
-        }
-        return processedDatasets;
-    },
-
-    getKeyOfObject: function (object, key, defaultValue) {
-        var returnVal = key.split(".").reduce(function (o, x) {
-            return (typeof o === 'undefined' || o === null) ? o : o[x];
-        }, object);
-
-        return typeof returnVal !== 'undefined' ? returnVal : defaultValue;
-    },
-
-    getTicksConfig: function (config, axis) {
-        var ticksConf = {
-            title: {text: this.getKeyOfObject(config, 'axis.' + axis + '.label', undefined)}
-        };
-        var calculateAutomatic = this.getKeyOfObject(config, 'axis.' + axis + '.auto', 0) === '1';
-        if (calculateAutomatic === true) {
-            return ticksConf;
-        }
-
-        var min = this.getKeyOfObject(config, 'axis.' + axis + '.min', null);
-        if (min !== null) {
-            ticksConf.min = parseInt(min, 10);
-        }
-        var max = this.getKeyOfObject(config, 'axis.' + axis + '.max', null);
-        if (max !== null && max > min) {
-            ticksConf.max = parseInt(max, 10);
-        }
-        return ticksConf;
-    },
-
-    getChartOptions: function (element) {
-        var chartConfig = {};
-        try {
-            chartConfig = JSON.parse(element.getAttribute('data-chart-config')) || {};
-        } catch (e) {
-        }
-
-        if (Object.keys(chartConfig).length <= 0) {
-            return {};
-        }
-        return chartConfig;
-    }
-};
-
-document.addEventListener("DOMContentLoaded", function () {
-    Hoogi91.Charts.init();
-});
+!function(){"use strict";var t={};t.default={chartsData:null,chartsContainer:null,init:function(){if("object"==typeof window["Hoogi91.chartsData"]&&(this.chartsData=window["Hoogi91.chartsData"]),this.chartsContainer=document.querySelectorAll(".chart-container > .chart"),this.chartsContainer.length>0&&null!==this.chartsData&&Object.keys(this.chartsData).length>0)for(var t=0;t<this.chartsContainer.length;++t){var e=this.chartsContainer[t],a=e.getAttribute("data-chart-type")||"",r=e.getAttribute("data-chart-data")||"";if(!(a.length<=0||r.length<=0)){var n=this.getChartData(r);if(!(n.length<=0)){var i=n.labels||[],o=n.datasets||[];if(i.length>0&&o.length>0)switch(a){case"chart_bar":this.createBarChart(e,i,o);break;case"chart_line":this.createLineChart(e,i,o);break;case"chart_pie":this.createPieChart(e,i,o);break;case"chart_doughnut":this.createDoughnutChart(e,i,o)}}}}},getChartData:function(t){return void 0===this.chartsData[t]?{}:this.chartsData[t]},createBarChart:function(t,e,a){var r=this,n=r.getChartOptions(t),i=new ApexCharts(t,{chart:{type:"bar",stacked:"1"===r.getKeyOfObject(n,"bar.stacked",0)},series:r.createDatasets(a,(function(t){var n=r.createDatasets(t.data,(function(r,n){var i={x:e[n]||"",y:r};return 1===a.length&&t.background.length>1&&(i.fillColor=t.background[n]||null),1===a.length&&t.border.length>1&&(i.strokeColor=t.border[n]||null),i}));return{name:t.label,data:n}})),labels:e,colors:a[0].border||a[0].background||void 0,fill:{colors:a[0].background||a[0].border||void 0},dataLabels:{enabled:!1},stroke:{width:1},plotOptions:{bar:{horizontal:"1"===r.getKeyOfObject(n,"bar.horizontal",0)}},xaxis:r.getTicksConfig(n,"x"),yaxis:r.getTicksConfig(n,"y"),legend:{show:"1"===r.getKeyOfObject(n,"legend.active",0),position:r.getKeyOfObject(n,"legend.position","top")}});return i.render(),i},createLineChart:function(t,e,a){var r=this,n=r.getChartOptions(t),i=new ApexCharts(t,{chart:{type:"1"===r.getKeyOfObject(n,"line.fill",0)?"area":"line",stacked:"1"===r.getKeyOfObject(n,"line.stacked",0)},series:r.createDatasets(a,(function(t){return{name:t.label,data:t.data}})),labels:e,colors:a[0].border||a[0].background||void 0,fill:{colors:a[0].background||a[0].border||void 0},dataLabels:{enabled:!1},stroke:{colors:a[0].background||void 0,curve:r.getKeyOfObject(n,"line.stepped","smooth")},markers:{size:[5,7],colors:a[0].border||void 0},xaxis:r.getTicksConfig(n,"x"),yaxis:r.getTicksConfig(n,"y"),legend:{show:"1"===r.getKeyOfObject(n,"legend.active",0),position:r.getKeyOfObject(n,"legend.position","top")}});return i.render(),i},createPieChart:function(t,e,a){var r=this,n=r.getChartOptions(t),i=new ApexCharts(t,{chart:{type:"pie"},colors:a[0].background||a[0].border||[],series:a[0].data||[],labels:e,dataLabels:{enabled:!1},legend:{show:"1"===r.getKeyOfObject(n,"legend.active",0),position:r.getKeyOfObject(n,"legend.position","top")}});return i.render(),i},createDoughnutChart:function(t,e,a){var r=this,n=r.getChartOptions(t),i=parseInt(r.getKeyOfObject(n,"doughnut.cutoutPercentage",75),10),o=new ApexCharts(t,{chart:{type:"donut"},colors:a[0].background||a[0].border||[],series:a[0].data||[],labels:e,dataLabels:{enabled:!1},plotOptions:{pie:{donut:{size:i<=0?75:i}}},legend:{show:"1"===r.getKeyOfObject(n,"legend.active",0),position:r.getKeyOfObject(n,"legend.position","top")}});return o.render(),o},createDatasets:function(t,e){var a=e||null,r=[];if("function"!=typeof a)return r;if("object"==typeof t&&t.length>0)for(var n=0;n<t.length;++n){var i=a(t[n],n);r.push(i)}return r},getKeyOfObject:function(t,e,a){var r=e.split(".").reduce((function(t,e){return null==t?t:t[e]}),t);return void 0!==r?r:a},getTicksConfig:function(t,e){var a={title:{text:this.getKeyOfObject(t,"axis."+e+".label",void 0)}};if(!0===("1"===this.getKeyOfObject(t,"axis."+e+".auto",0)))return a;var r=this.getKeyOfObject(t,"axis."+e+".min",null);null!==r&&(a.min=parseInt(r,10));var n=this.getKeyOfObject(t,"axis."+e+".max",null);return null!==n&&n>r&&(a.max=parseInt(n,10)),a},getChartOptions:function(t){var e={};try{e=JSON.parse(t.getAttribute("data-chart-config"))||{}}catch(t){}return Object.keys(e).length<=0?{}:e}},window["Hoogi91.Charts"]=t.default}();
