+++ Resources/Public/JavaScript/chartjs.js
@@ -1,338 +1,1 @@
-var Hoogi91 = Hoogi91 || {};
-Hoogi91.Charts = {
-    chartsData: null,
-    chartsContainer: null,
-
-    /**
-     * initialize all charts that can be found if available options are filled
-     */
-    init: function () {
-        if (typeof Hoogi91.chartsData === 'object') {
-            this.chartsData = Hoogi91.chartsData;
-        }
-
-        this.chartsContainer = document.querySelectorAll('.chart-container > .chart');
-        if (this.chartsContainer.length > 0 && this.chartsData !== null && Object.keys(this.chartsData).length > 0) {
-            for (var i = 0; i < this.chartsContainer.length; ++i) {
-                var element = this.chartsContainer[i];
-
-                // check id, type and datakey of current chart element and break if something is not available
-                var type = element.getAttribute('data-chart-type') || '';
-                var dataKey = element.getAttribute('data-chart-data') || '';
-                if (type.length <= 0 || dataKey.length <= 0) {
-                    continue;
-                }
-
-                var chartData = this.getChartData(dataKey);
-                if (chartData.length <= 0) {
-                    continue;
-                }
-
-                var labels = chartData.labels || [];
-                var datasets = chartData.datasets || [];
-                if (labels.length > 0 && datasets.length > 0) {
-                    switch (type) {
-                        case 'chart_bar':
-                            this.createBarChart(element, labels, datasets);
-                            break;
-                        case 'chart_line':
-                            this.createLineChart(element, labels, datasets);
-                            break;
-                        case 'chart_pie':
-                            this.createPieChart(element, labels, datasets);
-                            break;
-                        case 'chart_doughnut':
-                            this.createDoughnutChart(element, labels, datasets);
-                            break;
-                    }
-                }
-            }
-        }
-    },
-
-    /**
-     * get chart data by key
-     *
-     * @param key
-     * @returns object
-     */
-    getChartData: function (key) {
-        if (typeof this.chartsData[key] === 'undefined') {
-            return {};
-        }
-        return this.chartsData[key];
-    },
-
-    /**
-     * build bar chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createBarChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        return new Chart(element.getContext('2d'), {
-            type: 'bar',
-            data: {
-                labels: labels,
-                datasets: _this.createDatasets(datasets, function (set, index) {
-                    return {
-                        label: set['label'],
-                        data: set['data'],
-                        backgroundColor: datasets.length === 1
-                            ? set['background'] || []
-                            : set['background'][index] || [],
-                        borderColor: datasets.length === 1
-                            ? set['border'] || []
-                            : set['border'][index] || [],
-                        borderWidth: 1
-                    }
-                })
-            },
-            options: {
-                plugins: {
-                    legend: {
-                        display: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                        position: _this.getKeyOfObject(options, 'legend.position', 'top'),
-                    },
-                },
-                indexAxis: _this.getKeyOfObject(options, 'bar.horizontal', 0) === '1' ? 'y' : 'x',
-                scales: {
-                    x: {
-                        stacked: _this.getKeyOfObject(options, 'bar.stacked', 0) === '1',
-                        ticks: _this.getTicksConfig(options, 'x'),
-                        title: _this.getAxisLabelConfig(options, 'x'),
-                    },
-                    y: {
-                        stacked: _this.getKeyOfObject(options, 'bar.stacked', 0) === '1',
-                        ticks: _this.getTicksConfig(options, 'y'),
-                        title: _this.getAxisLabelConfig(options, 'y'),
-                    }
-                }
-            }
-        });
-    },
-
-    /**
-     * build line chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createLineChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        var steppedValue = _this.getKeyOfObject(options, 'line.stepped', false);
-        if (steppedValue !== 'after') {
-            steppedValue = steppedValue === '1';
-        }
-
-        var tensionValue = 0.4;
-        if (_this.getKeyOfObject(options, 'line.interpolation', 0) !== '1') {
-            tensionValue = 0;
-        }
-
-        return new Chart(element.getContext('2d'), {
-            type: 'line',
-            data: {
-                labels: labels,
-                datasets: _this.createDatasets(datasets, function (set, index) {
-                    return {
-                        label: set['label'],
-                        data: set['data'],
-                        fill: _this.getKeyOfObject(options, 'line.fill', 0) === '1',
-                        backgroundColor: datasets.length === 1
-                            ? set['border'] || []
-                            : set['border'][index] || [],
-                        borderColor: datasets.length === 1
-                            ? set['background'] || []
-                            : set['background'][index] || [],
-                    }
-                })
-            },
-            options: {
-                plugins: {
-                    legend: {
-                        display: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                        position: _this.getKeyOfObject(options, 'legend.position', 'top')
-                    },
-                },
-                scales: {
-                    x: {
-                        ticks: _this.getTicksConfig(options, 'x'),
-                        title: _this.getAxisLabelConfig(options, 'x'),
-                    },
-                    y: {
-                        stacked: _this.getKeyOfObject(options, 'line.stacked', 0) === '1',
-                        ticks: _this.getTicksConfig(options, 'y'),
-                        title: _this.getAxisLabelConfig(options, 'y'),
-                    }
-                },
-                elements: {
-                    line: {
-                        stepped: steppedValue,
-                        tension: tensionValue
-                    }
-                }
-            }
-        });
-    },
-
-    /**
-     * build pie chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createPieChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        return new Chart(element.getContext('2d'), {
-            type: 'pie',
-            data: {
-                labels: labels,
-                datasets: _this.createDatasets(datasets, function (set) {
-                    return {
-                        label: set['label'],
-                        data: set['data'],
-                        backgroundColor: typeof set['background'] === 'object' ? set['background'] : [],
-                        borderWidth: 1
-                    }
-                })
-            },
-            options: {
-                plugins: {
-                    legend: {
-                        display: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                        position: _this.getKeyOfObject(options, 'legend.position', 'top')
-                    },
-                },
-            }
-        });
-    },
-
-    /**
-     * build doughnut chart on element with labels and datasets
-     *
-     * @param element
-     * @param labels
-     * @param datasets
-     */
-    createDoughnutChart: function (element, labels, datasets) {
-        var _this = this;
-        var options = _this.getChartOptions(element);
-        var cutoutValue = parseInt(_this.getKeyOfObject(options, 'doughnut.cutoutPercentage', 75), 10);
-        cutoutValue = cutoutValue <= 0 ? 75 : cutoutValue;
-
-        return new Chart(element.getContext('2d'), {
-            type: 'doughnut',
-            data: {
-                labels: labels,
-                datasets: _this.createDatasets(datasets, function (set) {
-                    return {
-                        label: set['label'],
-                        data: set['data'],
-                        backgroundColor: typeof set['background'] === 'object' ? set['background'] : [],
-                        borderWidth: 1
-                    }
-                })
-            },
-            options: {
-                cutout: cutoutValue,
-                plugins: {
-                    legend: {
-                        display: _this.getKeyOfObject(options, 'legend.active', 0) === '1',
-                        position: _this.getKeyOfObject(options, 'legend.position', 'top')
-                    },
-                },
-            }
-        });
-    },
-
-    /**
-     * create datasets by given data and an additional mapping callback
-     *
-     * @param datasets
-     * @param mapping
-     * @returns array
-     */
-    createDatasets: function (datasets, mapping) {
-        var keyValueMapping = mapping || null,
-            processedDatasets = [];
-
-        if (typeof keyValueMapping !== 'function') {
-            return processedDatasets;
-        }
-
-        if (typeof datasets === 'object' && datasets.length > 0) {
-            for (var i = 0; i < datasets.length; ++i) {
-                var set = keyValueMapping(datasets[i], i);
-                processedDatasets.push(set);
-            }
-        }
-        return processedDatasets;
-    },
-
-    getKeyOfObject: function (object, key, defaultValue) {
-        var returnVal = key.split(".").reduce(function (o, x) {
-            return (typeof o === 'undefined' || o === null) ? o : o[x];
-        }, object);
-
-        return typeof returnVal !== 'undefined' ? returnVal : defaultValue;
-    },
-
-    getAxisLabelConfig: function (config, axis) {
-        var axisLabel = this.getKeyOfObject(config, 'axis.' + axis + '.label', '');
-        if (axisLabel.length > 0) {
-            return {
-                display: true,
-                text: axisLabel,
-            };
-        }
-        return {};
-    },
-
-    getTicksConfig: function (config, axis) {
-        var calculateAutomatic = this.getKeyOfObject(config, 'axis.' + axis + '.auto', 0) === '1';
-        if (calculateAutomatic === true) {
-            var stepSize = this.getKeyOfObject(config, 'axis.' + axis + '.step', null);
-            return stepSize !== null && stepSize > 0 ? {stepSize: parseInt(stepSize, 10)} : {};
-        }
-
-        var ticksConf = {};
-        var min = this.getKeyOfObject(config, 'axis.' + axis + '.min', null);
-        if (min !== null) {
-            ticksConf.min = parseInt(min, 10);
-        }
-        var max = this.getKeyOfObject(config, 'axis.' + axis + '.max', null);
-        if (max !== null && max > min) {
-            ticksConf.max = parseInt(max, 10);
-        }
-        var step = this.getKeyOfObject(config, 'axis.' + axis + '.step', null);
-        if (step !== null && step > 0) {
-            ticksConf.stepSize = parseInt(step, 10);
-        }
-        return ticksConf;
-    },
-
-    getChartOptions: function (element) {
-        var chartConfig = {};
-        try {
-            chartConfig = JSON.parse(element.getAttribute('data-chart-config')) || {};
-        } catch (e) {
-        }
-
-        if (Object.keys(chartConfig).length <= 0) {
-            return {};
-        }
-        return chartConfig;
-    }
-};
-
-document.addEventListener("DOMContentLoaded", function () {
-    Hoogi91.Charts.init();
-});
+!function(){"use strict";var t={};t.default={chartsData:null,chartsContainer:null,init:function(){if("object"==typeof window["Hoogi91.chartsData"]&&(this.chartsData=window["Hoogi91.chartsData"]),this.chartsContainer=document.querySelectorAll(".chart-container > .chart"),this.chartsContainer.length>0&&null!==this.chartsData&&Object.keys(this.chartsData).length>0)for(var t=0;t<this.chartsContainer.length;++t){var e=this.chartsContainer[t],a=e.getAttribute("data-chart-type")||"",n=e.getAttribute("data-chart-data")||"";if(!(a.length<=0||n.length<=0)){var r=this.getChartData(n);if(!(r.length<=0)){var i=r.labels||[],s=r.datasets||[];if(i.length>0&&s.length>0)switch(a){case"chart_bar":this.createBarChart(e,i,s);break;case"chart_line":this.createLineChart(e,i,s);break;case"chart_pie":this.createPieChart(e,i,s);break;case"chart_doughnut":this.createDoughnutChart(e,i,s)}}}}},getChartData:function(t){return void 0===this.chartsData[t]?{}:this.chartsData[t]},createBarChart:function(t,e,a){var n=this,r=n.getChartOptions(t);return new Chart(t.getContext("2d"),{type:"bar",data:{labels:e,datasets:n.createDatasets(a,(function(t,e){return{label:t.label,data:t.data,backgroundColor:1===a.length?t.background||[]:t.background[e]||[],borderColor:1===a.length?t.border||[]:t.border[e]||[],borderWidth:1}}))},options:{plugins:{legend:{display:"1"===n.getKeyOfObject(r,"legend.active",0),position:n.getKeyOfObject(r,"legend.position","top")}},indexAxis:"1"===n.getKeyOfObject(r,"bar.horizontal",0)?"y":"x",scales:{x:{stacked:"1"===n.getKeyOfObject(r,"bar.stacked",0),ticks:n.getTicksConfig(r,"x"),title:n.getAxisLabelConfig(r,"x")},y:{stacked:"1"===n.getKeyOfObject(r,"bar.stacked",0),ticks:n.getTicksConfig(r,"y"),title:n.getAxisLabelConfig(r,"y")}}}})},createLineChart:function(t,e,a){var n=this,r=n.getChartOptions(t),i=n.getKeyOfObject(r,"line.stepped",!1);"after"!==i&&(i="1"===i);var s=.4;return"1"!==n.getKeyOfObject(r,"line.interpolation",0)&&(s=0),new Chart(t.getContext("2d"),{type:"line",data:{labels:e,datasets:n.createDatasets(a,(function(t,e){return{label:t.label,data:t.data,fill:"1"===n.getKeyOfObject(r,"line.fill",0),backgroundColor:1===a.length?t.border||[]:t.border[e]||[],borderColor:1===a.length?t.background||[]:t.background[e]||[]}}))},options:{plugins:{legend:{display:"1"===n.getKeyOfObject(r,"legend.active",0),position:n.getKeyOfObject(r,"legend.position","top")}},scales:{x:{ticks:n.getTicksConfig(r,"x"),title:n.getAxisLabelConfig(r,"x")},y:{stacked:"1"===n.getKeyOfObject(r,"line.stacked",0),ticks:n.getTicksConfig(r,"y"),title:n.getAxisLabelConfig(r,"y")}},elements:{line:{stepped:i,tension:s}}}})},createPieChart:function(t,e,a){var n=this,r=n.getChartOptions(t);return new Chart(t.getContext("2d"),{type:"pie",data:{labels:e,datasets:n.createDatasets(a,(function(t){return{label:t.label,data:t.data,backgroundColor:"object"==typeof t.background?t.background:[],borderWidth:1}}))},options:{plugins:{legend:{display:"1"===n.getKeyOfObject(r,"legend.active",0),position:n.getKeyOfObject(r,"legend.position","top")}}}})},createDoughnutChart:function(t,e,a){var n=this,r=n.getChartOptions(t),i=parseInt(n.getKeyOfObject(r,"doughnut.cutoutPercentage",75),10);return i=i<=0?75:i,new Chart(t.getContext("2d"),{type:"doughnut",data:{labels:e,datasets:n.createDatasets(a,(function(t){return{label:t.label,data:t.data,backgroundColor:"object"==typeof t.background?t.background:[],borderWidth:1}}))},options:{cutout:i,plugins:{legend:{display:"1"===n.getKeyOfObject(r,"legend.active",0),position:n.getKeyOfObject(r,"legend.position","top")}}}})},createDatasets:function(t,e){var a=e||null,n=[];if("function"!=typeof a)return n;if("object"==typeof t&&t.length>0)for(var r=0;r<t.length;++r){var i=a(t[r],r);n.push(i)}return n},getKeyOfObject:function(t,e,a){var n=e.split(".").reduce((function(t,e){return null==t?t:t[e]}),t);return void 0!==n?n:a},getAxisLabelConfig:function(t,e){var a=this.getKeyOfObject(t,"axis."+e+".label","");return a.length>0?{display:!0,text:a}:{}},getTicksConfig:function(t,e){if(!0===("1"===this.getKeyOfObject(t,"axis."+e+".auto",0))){var a=this.getKeyOfObject(t,"axis."+e+".step",null);return null!==a&&a>0?{stepSize:parseInt(a,10)}:{}}var n={},r=this.getKeyOfObject(t,"axis."+e+".min",null);null!==r&&(n.min=parseInt(r,10));var i=this.getKeyOfObject(t,"axis."+e+".max",null);null!==i&&i>r&&(n.max=parseInt(i,10));var s=this.getKeyOfObject(t,"axis."+e+".step",null);return null!==s&&s>0&&(n.stepSize=parseInt(s,10)),n},getChartOptions:function(t){var e={};try{e=JSON.parse(t.getAttribute("data-chart-config"))||{}}catch(t){}return Object.keys(e).length<=0?{}:e}},window["Hoogi91.Charts"]=t.default}();
