+++ package.json
@@ -28,7 +28,7 @@
     "eslint": "^7.28.0",
     "eslint-loader": "^4.0.2",
     "sortablejs": "^1.14.0",
-    "terser-webpack-plugin": "^5.1.3",
+    "terser-webpack-plugin": "^5.3.1",
     "webpack": "^5.38.1",
     "webpack-cli": "^4.7.2"
   },
